Egg Dealer
Eatloog is an egg dealer, that sells, uh…wait for it………eggs! It is a supplier that targets restaurant owners, bakeries, and sari-sari store owners, since they tend to buy a lot. Each egg is Php 7, but buying a dozen eggs only costs Php 70.

Slogan: 12 eggs for the price of 10! Buyers are encouraged to buy whole trays of eggs for the savings. Eggs that don’t make a dozen are sold at the individual price of Php 7. Write a program that accepts the number of eggs as input, and prints the price as output.