Compound Interest
A compound interest, simply put, allows you to earn more money on your savings deposit! It is done by applying the interest rate on the current balance rather than the principal (initial amount). If you only deposit money once, say 1000 pesos on a 10% interest, your money will grow as follows:

Year 0: 1,000.00
Year 1: 1,100.00  (1,000 + 10%)
Year 2: 1,210.00  (1,100 + 10%)
Year 3: 1,331.00  (1,210 + 10%)
Year 4: 1,464.10  (1,331 + 10%, etc.)
Year 5: 1,610.51
The formula to achieve that is:

P * (1 + r)n

where P is the principal, r is the interest rate (e.g. 0.07 for 7%), and n is the number of years.

Write a program that accepts the principal, interest rate, and the number of years, and outputs the total amount in the savings account after the number of years input by the user.

Since it's money, format all amounts to 2 decimal places. Show 1005.10 rather than 1005.1, for example.

Input order: principal, interest rate, number of years