Hotel Tax and Tip
As the new normal arises, some people book hotels / inns for quarantine purposes. We all know that the hotel fees include accommodation stay, emergency fund for frontliners and TAX . Create a program that asks the user for the total amount paid and computes the tip(emergency fund for frontliners) as 10 percent of the accommodation amount (without the tax). Assume that this is in Philippine setup and that, tax is at 12 percent.

Format the output so that all of the values are displayed using two decimal places