FALL for you
Falling in love is human bliss and sorrow. When you love someone, you to have a 99% chance of getting hurt along the way. It's the same as the study of gravity. The more you distance yourself the less the force that is applied to you. When the object is falling, the height decreases and the attraction increases.

Create a program that determines how quickly an object is traveling as it falls to the ground from a given height. The program is given the distance as input. Also, assume that the acceleration due to gravity is 9.8 m/s2 and the initial velocity is 0 m/s.

Formula : speed.PNG

Such that,

vf is the final velocity or the velocity that the object is falling towards the ground
vi is the initial velocity (assumed to be zero (0 m/s))
a is accelaration due to gravity (9.8 m/s2)
d is the distance travelled (INPUT)
Note : Answers should be in two decimal places only.