Area of a polygon
Geometry is one of Lester's favorite subjects while he was studying. Now, he is one of the top-notch developers who develop tutorial apps for young kids to enjoy. One of its feature is a regular polygon area calculator.

Create a program that would assist Lester's users by creating a regular polygon area calculator. The program takes two numbers from the user: (1) side length and (2) number of sides.
Given the side length and number of sides, the formula for the area of a polygon is as follows:
Area of regular polygon = (number of sides * side length * side length) / 4 * tan(PI / number of sides)