Chinese Zodiac
The Chinese zodiac assigns animals to years in a 12 year cycle. One 12 year cycle is shown in the table below. The pattern is repeated with 2012 being another year of the dragon, and 1999 being another year of the hare.
Write a program that reads a year from the user and displays the animal associated with that year. 
Your program should work correctly for any year greater than or equal to zero, not just the ones listed in  the table.