Days in a month
The length of a month varies from 28 to 31 days. In this exercise you will create a program that reads the month number from the user. Then your program should display the number of days in that month.  Display “28 or 29 days” for an input of 2 (February) so that leap years are addressed.
Note: Please use modulo(%) as part of your solution.